add new steps 1-8

+++ src/main/java/com/jarzsoft/repository/FoclaasoRepository.java
@@ -1,6 +1,6 @@
 package com.jarzsoft.repository;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -11,8 +11,10 @@ import org.springframework.stereotype.Repository;
 import com.jarzsoft.entities.Foclaaso;
 
 @Repository
-public interface FoclaasoRepository extends JpaRepository<Foclaaso, BigDecimal> {
+public interface FoclaasoRepository extends JpaRepository<Foclaaso, BigInteger> {
 
 	@Query(value = "select foclaaso.* from foclaaso, fotabase_com where foclaaso.Activa = 'S' and  foclaaso.cod_inter  = fotabase_com.cod_inter and fotabase_com.codter_asesor = :usuario order by foclaaso.nom_claaso", nativeQuery = true)
 	List<Foclaaso> findByUsuario(@Param("usuario") String usuario);
+	
+
 }
